Fix mailer picking wrong release version
When creating the html version of the mail when a release is set, the
`helper_issues_show_detail_after_setting hook` is called twice. The
first call updates the hook context so that the value is set for the
chosen release, for example "1.12.0". The second time, the hook attempts
to find a release with id "1.12.0" which rails automatically turns into
"1" (since id is an integer). This causes the finder to find the release
that has id=1, in our case. "1.3.1", sending the incorrect version in
the mail.